Retrieve navbar to seperate navbar.js file
