+++ src/index.js
@@ -1,4 +1,12 @@
+import navbarLoad from './navbar';
 import pageLoad from './home';
 
+function component() {
+  const element = document.createElement('div');
+  element.appendChild(navbarLoad());
+  element.appendChild(pageLoad());
 
-document.body.appendChild(pageLoad());
+  return element;
+}
+
+document.body.appendChild(component());
